+++ src/main/java/org/folio/linked/data/configuration/json/deserialization/instance/MapRequestDeserializer.java
@@ -8,33 +8,33 @@ import static org.folio.ld.dictionary.ResourceTypeDictionary.ID_UNKNOWN;
 
 import java.util.Map;
 import org.folio.linked.data.domain.dto.IanField;
-import org.folio.linked.data.domain.dto.IdentifierField;
 import org.folio.linked.data.domain.dto.IsbnField;
 import org.folio.linked.data.domain.dto.IssnField;
 import org.folio.linked.data.domain.dto.LccnField;
+import org.folio.linked.data.domain.dto.MapRequest;
 import org.folio.linked.data.domain.dto.OtherIdField;
 import org.folio.linked.data.util.DtoDeserializer;
 import tools.jackson.core.JsonParser;
 import tools.jackson.databind.DeserializationContext;
 import tools.jackson.databind.ValueDeserializer;
 
-public class IdentifierFieldDeserializer extends ValueDeserializer<IdentifierField> {
+public class MapRequestDeserializer extends ValueDeserializer<MapRequest> {
 
-  private static final Map<String, Class<? extends IdentifierField>> IDENTITY_MAP = Map.of(
+  private static final Map<String, Class<? extends MapRequest>> IDENTITY_MAP = Map.of(
     ID_LCCN.getUri(), LccnField.class,
     ID_ISSN.getUri(), IssnField.class,
     ID_ISBN.getUri(), IsbnField.class,
     ID_IAN.getUri(), IanField.class,
     ID_UNKNOWN.getUri(), OtherIdField.class
   );
-  private final DtoDeserializer<IdentifierField> dtoDeserializer;
+  private final DtoDeserializer<MapRequest> dtoDeserializer;
 
-  public IdentifierFieldDeserializer() {
-    dtoDeserializer = new DtoDeserializer<>(IdentifierField.class, IDENTITY_MAP);
+  public MapRequestDeserializer() {
+    dtoDeserializer = new DtoDeserializer<>(MapRequest.class, IDENTITY_MAP);
   }
 
   @Override
-  public IdentifierField deserialize(JsonParser jp, DeserializationContext dc) {
+  public MapRequest deserialize(JsonParser jp, DeserializationContext dc) {
     return dtoDeserializer.deserialize(jp, dc);
   }
 }