+++ src/main/java/org/folio/linked/data/mapper/dto/resource/common/identifier/IanMapperUnit.java
@@ -1,4 +1,4 @@
-package org.folio.linked.data.mapper.dto.resource.common.instance.sub.identifier;
+package org.folio.linked.data.mapper.dto.resource.common.identifier;
 
 import static org.folio.ld.dictionary.PredicateDictionary.MAP;
 import static org.folio.ld.dictionary.ResourceTypeDictionary.ID_IAN;
@@ -6,9 +6,9 @@ import static org.folio.ld.dictionary.ResourceTypeDictionary.ID_IAN;
 import org.folio.ld.dictionary.ResourceTypeDictionary;
 import org.folio.linked.data.domain.dto.IanField;
 import org.folio.linked.data.domain.dto.IanFieldResponse;
-import org.folio.linked.data.domain.dto.IdentifierFieldResponse;
 import org.folio.linked.data.domain.dto.IdentifierRequest;
 import org.folio.linked.data.domain.dto.IdentifierResponse;
+import org.folio.linked.data.domain.dto.MapResponse;
 import org.folio.linked.data.mapper.dto.resource.base.CoreMapper;
 import org.folio.linked.data.mapper.dto.resource.base.MapperUnit;
 import org.folio.linked.data.service.resource.hash.HashService;
@@ -23,7 +23,7 @@ public class IanMapperUnit extends AbstractIdentifierMapperUnit {
   }
 
   @Override
-  protected IdentifierFieldResponse toFieldResponse(IdentifierResponse identifierResponse) {
+  protected MapResponse toFieldResponse(IdentifierResponse identifierResponse) {
     return new IanFieldResponse().ian(identifierResponse);
   }