+++ src/main/java/org/folio/linked/data/service/profile/ProfileService.java
@@ -1,6 +1,7 @@
 package org.folio.linked.data.service.profile;
 
 import java.util.List;
+import org.folio.ld.dictionary.ResourceTypeDictionary;
 import org.folio.linked.data.domain.dto.ProfileMetadata;
 import org.folio.linked.data.model.entity.Profile;
 
@@ -10,4 +11,6 @@ public interface ProfileService {
   Profile getProfileById(Integer id);
 
   List<ProfileMetadata> getMetadataByResourceType(String resourceTypeUri);
+
+  ResourceTypeDictionary getResourceTypeByProfileId(Integer profileId);
 }