+++ src/main/java/org/folio/linked/data/util/Constants.java
@@ -29,6 +29,7 @@ public class Constants {
     public static final String SETTINGS_ENTRIES = "settings-entries";
     public static final String MODULE_STATE = "module-state";
     public static final String PROFILES = "profiles";
+    public static final String PROFILES_RESOURCE_TYPE = "profiles-resource-type";
     public static final String AUTHORITY_SOURCE_FILES = "authority-source-files";
     public static final String VOCABULARIES = "vocabularies";
   }